Fix backport of "netfilter: x_tables: validate targets of jumps"
In 3.16, xt_table_info::entries points to an array of per-CPU pointers
to filter tables, not to a single filter table.  So find_jump_target()
was looking at the wrong data.

Use the version that went into 3.14.73, which makes the new filter
table a separate parameter to find_jump_target().