fs: Fix oops when fcntl() is called on an aufs directory
CVE-2016-7118; regression in 3.16.36-1.

In modifying the check for the setfl operation to be safe with old
modules that have a smaller file_operations structure, I mistakenly
dropped the check that the operation pointer is non-null.  The
modified check rejects file_operations in any module other than the
rebuilt aufs, but it is still necessary to check the pointer because
for aufs directories it is NULL.