Revert "random: fix crng_ready() test"
Some systems (particularly VMs, but also bare metal installations) are
not getting enough entropy to boot successfully with this change.  We
need to either improve the amount of entropy available or change the
affected applications to be more tolerant of this situation.  Defer
this fix until we decide how to avoid the regression.

Closes: #897599, but reopens CVE-2018-1108.