[sh4] Delete ineffective setting of CONFIG_PREEMPT_NONE
If multiple options in a "choice" group are enabled, kconfig uses the
last (and warns about this).  This means that the global default of
CONFIG_PREEMPT_VOLUNTARY=y has always overridden the setting of
CONFIG_PREEMPT_NONE=y here.

Since no-one seems to have complained, assume that the default is OK
and remove this.