[x86] sound/soc/intel: Explicitly enable SND_SST_ATOM_HIFI2_PLATFORM_ACPI as module
The HiFi2 driver was split up into ACPI and PCI parts, and we need to
select one or both of those instead of SND_SST_ATOM_HIFI2_PLATFORM.
Baed on the help text, it appears that the ACPI part is all we want.
It also gets selected automatically, so this commit doesn't change
the resulting config.