[x86] Enable TI TPS6598x USB Power Delivery controller family
This enables USB power delivery controller that is found on some
ThinkPad laptops, like the ThinkPad L380. This solves the following
error in i2c_hid:

[ 1.917100] i2c_hid i2c-INT3515:02: unexpected HID descriptor bcdVersion (0x0000)

This has the effect to use a dummy_regulator by default (no support is
found for INT3515, so vdd cannot be found) and disable pm_runtime during
probe.