[x86] mm: Consider effective protection attributes in W+X check
Upgrading for 4.9.130 to 4.9.135 introduced a false positive for
this check.  This was already fixed upstream, so pick that fix.

This conflicts with "[amd64] Don't WARN about expected W+X pages on
Xen", so move the latter afterward and refresh it.  (I also tried
replacing it with the proper fix, commit 2cc42bac1c "x86-64/Xen:
eliminate W+X mappings", but that resulted in a crash at boot.)