[powerpc*] udeb: Remove redundant lines from scsi-modules
The a100u2w, ibmvscsi, mac53c94, mesh, and ps3rom drivers are built
under drivers/scsi so don't need to be specifically included.

The ps3rom driver then doesn't need to be excluded on ppc64el.

The ps3disk driver is a SATA driver not a SCSI driver, so add a
comment to explain why it's here.