Update to 4.9.172
Drop Revert-block-loop-Use-global-lock-for-ioctl-operatio.patch

Drop "kbuild: Use -nostdinc in compile tests", which is no longer needed

This was a workaround for bug #717557, which was fixed long ago.  It
would no longer work around that bug anyway, because it doesn't touch
the compiler feature tests in scripts/Kconfig.include.

Cleanup debian/changelog file