[arm64] udeb: usb-modules: Include phy-rockchip-typec, extcon-usbc-cros-ec
On Samsung Chromebook Plus (v1) trying to boot from a rootfs on a USB
storage device without these modules in the initramfs, it drops to an
initramfs shell with a non-working display. For the d-i netboot image,
the screen doesn't turn on, but the installer menu works.

A recent change to initramfs-tools includes extcon-usbc-cros-ec, so
include that and a relevant PHY module here as well.

Relevant:
  https://salsa.debian.org/kernel-team/initramfs-tools/commit/994d698a

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>