[x86] iommu: Enable INTEL_IOMMU_DEFAULT_ON (Closes: #934309)
This option has been disabled since the Intel IOMMU driver was first
enabled in 2009 (commit da261d6b5896 "x86 enable DMAR").  This was
probably sensible back when the driver was new and didn't have
workarounds for some of the compatibility issues with integrated
GPUs.

At this point it should be safe to enable by default.  It is also
necessary (but not sufficient) to block "Thunderclap" type attacks.