debian/rules: Make maintainerclean delete (almost) everything clean does
The maintainerclean target should delete everything that doesn't
belong in a source package and can't be regenerated using the orig
and debian/control-real targets.  Given a git checkout, "git clean
-d -f -x" will do this, but not everyone builds from git.

Currently it doesn't delete everything that the clean target does, by
a long way.  Unlike the clean target it does not require
debian/control to exist or be buildable, so it cannot depend on clean
or invoke dh_clean.

Therefore:

* Move the file patterns from the clean target into a variable, and
  use that in both clean and maintainerclean targets
* Delete debhelper temporary files and debian/files in the
  maintainerclean target

This doesn't cover editor temporary/backup files, which dh_clean
removes, but they're less of an issue since dpkg-source excludes those
automatically.