debian/lib/python/debian_linux/abi.py: Add support for namespaces
Exported symbols can now be defined to belong to a specific namespace,
and Module.symvers includes this as an additional field between name
and module.

The namespace can be an empty string, so when reading we need to split
fields on '\t' and not the default of one-or-more-whitespace.  We then
also need to separate fields with '\t' when writing an ABI reference.

Namespaces are intended for grouping symbols exported for use by
groups of in-tree modules, and we ought to add support for ignoring
ABI changes on this basis.  For now, just add it as an attribute of
Symbol which is compared when checking for ABI changes.