signing_templates/rules.real: Include modules.builtin.modinfo file in image
Since 898490c010b5 ("moduleparam: Save information about built-in
modules in separate file") in 5.2-rc1, there is a new
modules.builtin.modinfo which contains modinfo from all modules that are
built into the kernel.

Whilst it was included into to unsigned packages, it was missing from
the signed packages recieved from the signing service as the file was
not included via the install-signed target.

Together with a kmod update to a snapshot release from 26+20191223-1,
depmod will generate an empty modules.builtin.alias.bin.

Closes: #948427
Link: https://bugs.debian.org/948257#88
Reported-by: Benjamin Poirier <benjamin.poirier@gmail.com>
Signed-off-by: Salvatore Bonaccorso <carnil@debian.org>