control.image.meta.in: set Provides: wireguard-modules
With wireguard now being part of Debian's kernel, it makes sense to have
the kernel package satisfy the virtual package slot, "wireguard-modules",
that is required by the wireguard meta package. This, notably, is what's
already done by Ubuntu's kernel as well as Arch's, and other distro
kernels are likely to follow. Note that this is a somewhat temporary
arrangement -- we only need this while we're keeping wireguard-dkms
around to support kernels that don't come with wireguard built-in.
Over time that number will dwindle, at which point we can get rid of
wireguard-dkms as well as reverting this commit. In other words, adding
this Provides: line is a temporary stop-gap solution while wireguard
transitions from being an out-of-tree module to an in-tree module.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>