Enable support for fsverity
fs-verity is a Linux kernel feature that does transparent on-demand
integrity/authenticity verification of the contents of read-only files,
using a hidden Merkle tree (hash tree) associated with the file. The
mechanism is similar to dm-verity, but implemented at the file level
rather than at the block device level. This enables support for both
integrity and authenticity.