Provide linux-{image,headers}-generic virtual packages on most architectures
The Debian installer can choose a suitable kernel to install based on
hardware information, but image builders can't do that and they need
to know a sensible default kernel for each architecture.

I've resisted defining such a default before now, because there is no
reasonable default on some architectures.  But we can still provide a
default for those architectures where there is:

* [alpha,sparc64] default to SMP flavour; both flavours should run on
  everything but this seems more useful
* [amd64] default to amd64, which should run on everything
* [arm64] default to arm64, which should run on everything
* [armhf] default to armmp, which should run on everything; non-LPAE
  CPUs are very common
* [hppa] default to parisc, which should run on everything
* [i386] default to 686-pae, which requires PAE; non-PAE CPUs are
  quite rare
* [ia64] default to itanium, which should run on everything
* [m68k,mips64r6*,ppc64*,riscv64,s390x] default to the only flavour
* [mipsr6*] default to 32-bit flavour; 64-bit systems should use the
  64-bit architecture

The architectures with no default are:

* [armel,mips,mips64,mips64el,mipsel,sh4] flavours are platform-specific
* [powerpc] 32-bit flavour won't boot on 64-bit systems and vice versa

Closes: #960209