udeb: Move any compression modules to kernel-image; drop compress-modules
With the current configuration on amd64, all of the compression
libraries are built-in and compress-modules ends up empty.

This may not be the case on all architectures and flavours, but I
don't want to have to keep track of which do and don't need a
compress-modules package.  Move the modules, if there are any, into
kernel-image, and drop the compress-modules packages entirely.