[x86] debug: Keep FUNCTION_ERROR_INJECTION enabled
CONFIG_FUNCTION_ERROR_INJECTION used to be automatic and always
enabled (where its dependencies were met), but with 4.19.268 it became
configurable and disabled by default.

Keep it enabled for buster.  Although this commit touches
debian/config/config, the dependencies are not met on arm64 or armhf
so this is specific to x86 in practice.